Crop_Project/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.0
matplotlib==3.7.2
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
reportlab==4.0.4
python-multipart==0.0.6
werkzeug==2.3.7